+++ src/js/components/autocomplete.js
@@ -64,9 +64,11 @@
             var $this   = this,
                 select  = false,
                 trigger = UI.Utils.debounce(function(e) {
-                    if(select) {
+
+                    if (select) {
                         return (select = false);
                     }
+
                     $this.handle();
                 }, this.options.delay);
 
@@ -87,9 +89,10 @@
             this.dropdown.attr('aria-expanded', 'false');
 
             this.input.on({
-                "keydown": function(e) {
 
-                    if (e && e.which && !e.shiftKey) {
+                keydown: function(e) {
+
+                    if (e && e.which && !e.shiftKey && $this.visible) {
 
                         switch (e.which) {
                             case 13: // enter
@@ -118,7 +121,8 @@
                     }
 
                 },
-                "keyup": trigger
+                
+                keyup: trigger
             });
 
             this.dropdown.on("click", ".uk-autocomplete-results > *", function(){
@@ -208,7 +212,9 @@
         },
 
         show: function() {
+
             if (this.visible) return;
+
             this.visible = true;
             this.element.addClass("uk-open");