+++ src/js/components/grid.js
@@ -24,7 +24,8 @@
             duration  : 300,
             gutter    : 0,
             controls  : false,
-            filter    : false
+            filter    : false,
+            origin    : UI.langdirection
         },
 
         boot:  function() {
@@ -54,6 +55,7 @@
             this.element.css({'position': 'relative'});
 
             this.controls = null;
+            this.origin   = this.options.origin;
 
             if (this.options.controls) {
 
@@ -78,23 +80,23 @@
                 if ($this.currentfilter) {
                     $this.filter($this.currentfilter);
                 } else {
-                    this.updateLayout();
+                    this.update();
                 }
 
             }.bind(this), 100));
 
             this.on('display.uk.check', function(){
-                if ($this.element.is(":visible"))  $this.updateLayout();
+                if ($this.element.is(":visible"))  $this.update();
             });
 
             UI.domObserve(this.element, function(e) {
-                $this.updateLayout();
+                $this.update();
             });
 
             if (this.options.filter !== false) {
                 this.filter(this.options.filter);
             } else {
-                this.updateLayout();
+                this.update();
             }
         },
 
@@ -115,16 +117,18 @@
 
             if (this.options.gutter) {
 
-                css['padding-left']   = this.gutterh;
+                css['padding-'+this.origin] = this.gutterh;
                 css['padding-bottom'] = this.gutterv;
 
-                this.element.css('margin-left', this.gutterh * -1);
+                this.element.css('margin-'+this.origin, this.gutterh * -1);
             }
 
             children.attr('data-grid-prepared', 'true').css(css);
         },
 
-        updateLayout: function(elements) {
+        update: function(elements) {
+
+            var $this = this;
 
             this._prepareElements();
 
@@ -136,7 +140,7 @@
                 top       = 0,
                 positions = [],
 
-                item, width, height, pos, i, z, max, size;
+                item, width, height, pos, posi, i, z, max, size;
 
             this.trigger('beforeupdate.uk.grid', [children]);
 
@@ -159,18 +163,21 @@
                     if (top <= pos.aY) { top = pos.aY; }
                 }
 
-                positions.push({
+                posi = {
                     "ele"    : item,
                     "top"    : top,
-                    "left"   : left,
                     "width"  : width,
                     "height" : height,
                     "aY"     : (top  + height),
                     "aX"     : (left + width)
-                });
+                };
+
+                posi[$this.origin] = left;
+
+                positions.push(posi);
             });
 
-            var posPrev, maxHeight = 0;
+            var posPrev, maxHeight = 0, positionto;
 
             // fix top
             for (i=0,max=positions.length;i<max;i++) {
@@ -183,7 +190,7 @@
                     posPrev = positions[z];
 
                     // (posPrev.left + 1) fixex 1px bug when using % based widths
-                    if (pos.left < posPrev.aX && (posPrev.left +1) < pos.aX) {
+                    if (pos[this.origin] < posPrev.aX && (posPrev[this.origin] +1) < pos.aX) {
                         top = posPrev.aY;
                     }
                 }
@@ -201,7 +208,11 @@
                 this.element.stop().animate({'height': maxHeight}, 100);
 
                 positions.forEach(function(pos){
-                    pos.ele.stop().animate({"top": pos.top, "left": pos.left, opacity: 1}, this.options.duration);
+
+                    positionto = {"top": pos.top, opacity: 1};
+                    positionto[$this.origin] = pos[$this.origin];
+
+                    pos.ele.stop().animate(positionto, this.options.duration);
                 }.bind(this));
 
             } else {
@@ -209,7 +220,9 @@
                 this.element.css('height', maxHeight);
 
                 positions.forEach(function(pos){
-                    pos.ele.css({"top": pos.top, "left": pos.left, opacity: 1});
+                    positionto = {"top": pos.top, opacity: 1};
+                    positionto[$this.origin] = pos[$this.origin];
+                    pos.ele.css(positionto);
                 }.bind(this));
             }
 
@@ -260,7 +273,7 @@
             elements.hidden.attr('aria-hidden', 'true').filter(':visible').fadeOut(this.options.duration);
             elements.visible.attr('aria-hidden', 'false').filter(':hidden').css('opacity', 0).show();
 
-            $this.updateLayout(elements.visible);
+            $this.update(elements.visible);
 
             if (this.controls && this.controls.length) {
                 this.controls.find('[data-uk-filter]').removeClass('uk-active').filter('[data-uk-filter="'+filter+'"]').addClass('uk-active');
@@ -287,7 +300,7 @@
 
             }).appendTo(this.element);
 
-            this.updateLayout(elements.filter(':visible'));
+            this.update(elements.filter(':visible'));
 
             if (this.controls && this.controls.length) {
                 this.controls.find('[data-uk-sort]').removeClass('uk-active').filter('[data-uk-sort="'+by+':'+(order == -1 ? 'desc':'asc')+'"]').addClass('uk-active');