+++ src/js/components/tooltip.js
@@ -21,13 +21,13 @@
     UI.component('tooltip', {
 
         defaults: {
-            "offset": 5,
-            "pos": "top",
-            "animation": false,
-            "delay": 0, // in miliseconds
-            "cls": "",
-            "activeClass": "uk-active",
-            "src": function(ele) {
+            offset: 5,
+            pos: 'top',
+            animation: false,
+            delay: 0, // in miliseconds
+            cls: "",
+            activeClass: "uk-active",
+            src: function(ele) {
                 var title = ele.attr('title');
 
                 if (title !== undefined) {